refactor/#307: RealityKit 엔티티 머테리얼 세팅 책임 분리
